refactor(state): extract context to context.rs
Move Context, ContextBuilder, ChainParams and ChainType enum to
dedicated context.rs module.
